+++ src/ts/PlotBoilerplate.ts
@@ -58,6 +58,7 @@
  * @modified 2020-07-06 Replacing Touchy.js by AlloyFinger.js
  * @modified 2020-07-27 Added the getVertexNear(XYCoords,number) function
  * @modified 2020-07-27 Extended the remove(Drawable) function: vertices are now removed, too.
+ * @modified 2020-07-28 Added PlotBoilerplate.revertMousePosition(number,number) –  the inverse function of transformMousePosition(...).
  * @version  1.9.0
  *
  * @file PlotBoilerplate
@@ -736,9 +737,9 @@ export class PlotBoilerplate {
 			    this.removeVertex( drawable.bezierCurves[i].startPoint, false );
 			    this.removeVertex( drawable.bezierCurves[i].startControlPoint, false );
 			    this.removeVertex( drawable.bezierCurves[i].endControlPoint, false );
-			    //if( i+1 == drawable.bezierCurves.length ) {
+			    if( i+1 == drawable.bezierCurves.length ) {
 				this.removeVertex( drawable.bezierCurves[i].endPoint, false );
-			    //}
+			    }
 			}
 		    } else if( drawable instanceof PBImage ) {
 			this.removeVertex( drawable.upperLeft, false );
@@ -765,7 +766,6 @@ export class PlotBoilerplate {
      * @return {void}
      **/
     removeVertex( vert:Vertex, redraw?:boolean ) : void {
-	// for( var i in this.drawables ) {
 	for( var i = 0; i < this.vertices.length; i++ ) {
 	    if( this.vertices[i] === vert ) {
 		this.vertices.splice(i,1);
@@ -778,11 +778,17 @@ export class PlotBoilerplate {
 
     /**
      * Find the vertex near the given position.
+     *
+     * The position is the absolute vertex position, not the x-y-coordinates on the canvas.
+     *
+     * @param {XYCoords} position - The position of the vertex to search for.
+     * @param {number} pixelTolerance - A radius around the position to include into the search. 
+     *                                  Note that the tolerance will be scaled up/down when zoomed.
+     * @return The vertex near the given position or undefined if none was found there.
      **/
-    getVertexNear( position:XYCoords, tolerance:number ) : Vertex|undefined {
-	var p : IDraggable|undefined = this.locatePointNear( position, // this.transformMousePosition(position.x, position.y),
-							     tolerance/Math.min(this.config.cssScaleX,this.config.cssScaleY) );
-	console.log(p);
+    getVertexNear( pixelPosition:XYCoords, pixelTolerance:number ) : Vertex|undefined {
+	var p : IDraggable|undefined = this.locatePointNear( this.transformMousePosition(pixelPosition.x, pixelPosition.y),
+							     pixelTolerance/Math.min(this.config.cssScaleX,this.config.cssScaleY) );
 	if( p && p.typeName == "vertex" )
 	    return this.vertices[p.vindex];
 	return undefined;
@@ -1283,12 +1289,26 @@ export class PlotBoilerplate {
      * @param {number} y - The y position relative to the canvas.
      * @instance
      * @memberof PlotBoilerplate
-     * @return {object} A simple object <pre>{ x : Number, y : Number }</pre> with the transformed coordinates.
+     * @return {XYCoords} A simple object <pre>{ x : Number, y : Number }</pre> with the transformed coordinates.
      **/
     transformMousePosition( x:number, y:number ) : XYCoords {
 	return { x : (x/this.config.cssScaleX-this.config.offsetX)/(this.config.scaleX),
 		 y : (y/this.config.cssScaleY-this.config.offsetY)/(this.config.scaleY) };
     };
+
+    /**
+     * Revert a transformed mouse position back to canvas coordinates.
+     *
+     * This is the inverse function of `transformMousePosition`.
+     *
+     * @param {number} x - The x component of the position to revert.
+     * @param {number} y - The y component of the position to revert.
+     * @return {XYCoords} The canvas coordinates for the given position.
+     **/
+    revertMousePosition( x:number, y:number ) : XYCoords {
+	return { x : x / this.config.cssScaleX + this.config.offsetX,
+		 y : y / this.config.cssScaleY + this.config.offsetY };
+    }